backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests
bs4
pandas
firebase-admin
geopy